[TAN-515] Move granular permissions into core - app & routes
