+++ back/app/policies/initiative_comment_policy.rb
@@ -56,7 +56,7 @@ class InitiativeCommentPolicy < ApplicationPolicy
   private
 
   def commenting_allowed?(user)
-    user # signed-in users can comment
+    !PermissionsService.new.denied_reason_for_resource user, 'commenting_initiative'
   end
 
   def owner?
