+++ back/app/serializers/web_api/v1/phase_serializer.rb
@@ -19,6 +19,8 @@ class WebApi::V1::PhaseSerializer < WebApi::V1::ParticipationContextSerializer
     user_basket object, params
   end
 
+  has_many :permissions
+
   def self.user_basket(object, params)
     preloaded_user_basket = params.dig(:user_baskets, object.id)&.first
     preloaded_user_basket || current_user(params)&.baskets&.select do |basket|
@@ -26,5 +28,3 @@ class WebApi::V1::PhaseSerializer < WebApi::V1::ParticipationContextSerializer
     end&.first
   end
 end
-
-WebApi::V1::PhaseSerializer.include(GranularPermissions::WebApi::V1::Patches::PhaseSerializer)
