--- back/engines/commercial/granular_permissions/app/policies/granular_permissions/patches/comment_reaction_policy.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module GranularPermissions
-  module Patches
-    module CommentReactionPolicy
-      def denied_for_initiative_reason(user)
-        PermissionsService.new.denied_reason_for_resource user, 'commenting_initiative'
-      end
-    end
-  end
-end
