--- back/engines/commercial/granular_permissions/app/policies/granular_permissions/patches/initiative_policy.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module GranularPermissions
-  module Patches
-    module InitiativePolicy
-      def posting_denied_reason(user)
-        PermissionsService.new.denied_reason_for_resource user, 'posting_initiative'
-      end
-    end
-  end
-end
