--- back/engines/commercial/granular_permissions/app/policies/granular_permissions/patches/initiative_reaction_policy.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module GranularPermissions
-  module Patches
-    module InitiativeReactionPolicy
-      def reacting_denied_reason(user)
-        PermissionsService.new.denied_reason_for_resource user, 'reacting_initiative'
-      end
-    end
-  end
-end
