--- back/engines/commercial/granular_permissions/app/serializers/granular_permissions/web_api/v1/patches/phase_serializer.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module GranularPermissions
-  module WebApi
-    module V1
-      module Patches
-        module PhaseSerializer
-          def self.included(base)
-            base.class_eval { has_many :permissions }
-          end
-        end
-      end
-    end
-  end
-end
