--- back/engines/commercial/granular_permissions/app/services/granular_permissions/patches/side_fx_participation_context_service.rb
@@ -1,27 +1,0 @@
-# frozen_string_literal: true
-
-module GranularPermissions
-  module Patches
-    module SideFxParticipationContextService
-      def self.prepended(base)
-        base.class_eval do
-          attr_writer :permissions_service
-        end
-      end
-
-      def after_create(pc, _user)
-        permissions_service.update_permissions_for_scope(pc)
-        super
-      end
-
-      def after_update(pc, _user)
-        permissions_service.update_permissions_for_scope(pc)
-        super
-      end
-
-      def permissions_service
-        @permissions_service ||= PermissionsService.new
-      end
-    end
-  end
-end
